Split transactions module to multiple files
