+++ src/store/transactions/index.js
@@ -1,0 +1,16 @@
+import actions from './actions';
+import getters from './getters';
+import mutations from './mutations';
+
+const state = {
+  pendingTransactions: [],
+  transactionHistory: [],
+};
+
+export default {
+  namespaced: true,
+  state,
+  actions,
+  getters,
+  mutations,
+};
